Added more unit tests. ShoppingCart is not covered yet

+++ src/main/java/com/javaguru/shoppinglist/domain/Product.java
@@ -48,7 +48,7 @@ public class Product {
         return discount;
     }
 
-    public void setDiscount(int discount) {
+    public void setDiscount(double discount) {
         this.discount = discount;
     }
